fix: now the articles have better format
