+++ package.json
@@ -47,10 +47,12 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
+    "dompurify": "^3.2.5",
     "embla-carousel-react": "^8.6.0",
     "framer-motion": "^12.9.2",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.503.0",
+    "marked": "^5.1.1",
     "next": "^15.3.1",
     "next-themes": "^0.4.6",
     "react": "^18.2.0",
